feat: split recipe/_api.py and _analysis.py into focused submodules
Splits _api.py into _recipe_ingredients.py (TypedDicts + formatting) and
_recipe_composition.py (sub-recipe merging). Splits _analysis.py into
_analysis_graph.py, _analysis_bfs.py, _analysis_blocks.py, and
_analysis_detectors.py. All external import paths remain stable via
re-exports in the reduced orchestrator files.
