+++ src/autoskillit/recipe/__init__.py
@@ -32,15 +32,17 @@ from autoskillit.recipe import rules_tools as _rules_tools  # noqa: E402 F401
 from autoskillit.recipe import rules_verdict as _rules_verdict  # noqa: E402 F401
 from autoskillit.recipe import rules_worktree as _rules_worktree  # noqa: E402 F401
 from autoskillit.recipe._api import (  # noqa: E402
+    format_recipe_list_response,
+    list_all,
+    load_and_validate,
+    validate_from_path,
+)
+from autoskillit.recipe._recipe_ingredients import (  # noqa: E402
     ListRecipesResult,
     LoadRecipeResult,
     RecipeListItem,
     build_ingredient_rows,
     format_ingredients_table,
-    format_recipe_list_response,
-    list_all,
-    load_and_validate,
-    validate_from_path,
 )
 from autoskillit.recipe.contracts import (  # noqa: E402
     StaleItem,
